+++ src/oboskrnl/scheduler/thread.h
@@ -66,7 +66,7 @@ typedef enum
 #if defined(__x86_64__)
 typedef __uint128_t thread_affinity;
 #elif defined(__m68k__)
-typedef uint64_t thread_affinity;
+typedef uint32_t thread_affinity;
 #endif
 extern OBOS_EXPORT thread_affinity Core_DefaultThreadAffinity;
 extern const uint64_t Core_ThreadPriorityToQuantum[THREAD_PRIORITY_MAX_VALUE+1];
